Move RUMInternalProxy onto the RUM object
Doing this prevents race conditions caused by lazy vars and simplifies initialziation.
