+++ Sources/Datadog/RUM/_RUMInternalProxy.swift
@@ -14,18 +14,16 @@ import Foundation
 ///
 /// Methods, members, and functionality of this class  are subject to change without notice, as they
 /// are not considered part of the public interface of the Datadog SDK.
-public class _InternalProxy {
-    let _telemtry = _TelemetryProxy()
-}
+public class _RUMInternalProxy {
+    weak var subscriber: RUMCommandSubscriber?
 
-public class _TelemetryProxy {
-    /// See Telementry.debug
-    func debug(id: String, message: String) {
-        DD.telemetry.debug(id: id, message: message)
+    init(subscriber: RUMCommandSubscriber) {
+        self.subscriber = subscriber
     }
 
-    /// See Telementry.error
-    func error(id: String, message: String, kind: String?, stack: String?) {
-        DD.telemetry.error(id: id, message: message, kind: kind, stack: stack)
+    func addLongTask(at: Date, duration: TimeInterval, attributes: [AttributeKey: AttributeValue] = [:]) {
+        let longTaskCommand = RUMAddLongTaskCommand(time: at, attributes: attributes, duration: duration)
+
+        subscriber?.process(command: longTaskCommand)
     }
 }
