+++ Sources/Datadog/RUMMonitor.swift
@@ -201,6 +201,8 @@ public class RUMMonitor: DDRUMMonitor, RUMCommandSubscriber {
 
         super.init()
 
+        _internalProxy = _RUMInternalProxy(subscriber: self)
+
         if Datadog.debugRUM {
             self.enableRUMDebugging(true)
         }
@@ -581,6 +583,12 @@ public class RUMMonitor: DDRUMMonitor, RUMCommandSubscriber {
         queue.async {
             self.rumAttributes[key] = nil
         }
+    }
+
+    // MARK: - Internal Proxy
+    var _internalProxy: _RUMInternalProxy?
+    override public var _internal: _RUMInternalProxy? {
+        get { return _internalProxy }
     }
 
     // MARK: - Internal
